Created resource generator
This allows for patches of resources to be placed. The "rarity" value of the resource determines how likely a new resource tile will spawn next to the starting point. The "spread" value determines how likely there is to be a new starting point next to the original for the patch to keep growing.